# Optimize file reads in `fs` module