--- a/Cargo.toml
+++ b/Cargo.toml
@@ -67,6 +67,8 @@ compression-gzip = ["async-compression/deflate", "async-compression/gzip"]
 [profile.release]
 codegen-units = 1
 incremental = false
+lto = "fat"
+strip = true
 
 [profile.bench]
 codegen-units = 1
@@ -98,7 +100,9 @@ required-features = ["websocket"]
 [[example]]
 name = "query_string"
 
-
 [[example]]
 name = "multipart"
 required-features = ["multipart"]
+
+[[example]]
+name = "file"